Ditched replay block. Long live AppendBlock
